230541108_OmerBAL/anomali_simulasyonu/soğutma-anomalisi-sim/requirements.txt
python-can==4.2.2
pytest==7.4.3
pyyaml==6.0.1
aiohttp==3.9.1
websockets==12.0
ocpp==0.14.1
cantools==2024.1.0